Add KeychainHelper to bundle KeyChain calls
